+++ kinetic/__init__.py
@@ -18,7 +18,7 @@
 
 # Protocol version
 protocol_version = '2.0.3'
-version = '0.7.0'
+version = '0.7.1'
 
 #utils
 from utils import buildRange