+++ CHANGES.md
@@ -6,6 +6,7 @@ This section will document changes to the library since the last release
 - Added getVersion and getVersionAsync to the library.
 
 ## Bug Fixes
+- Fixed tests not running and testcases with hardcoded 'localhost'
 - Fixed flush operation build parameters (Merge #5, contributed by @rpcope1).
 - AsyncClient returns NotConnected exception when an operation is attempted on a client before calling connect().
 - Lowered default number of keys asked on ranges to 200 (ASKOVAD-287).